refactor(maven): split parser into nested modules
Group the POM, manifest, properties, and test surfaces under src/parsers/maven so the large Maven parser follows the ADR-backed nested parser structure and is easier to extend safely.
