+++ src/parsers/golden_test.rs
@@ -79,7 +79,7 @@ mod helm_golden_test;
 mod hex_lock_golden_test;
 #[path = "julia_golden_test.rs"]
 mod julia_golden_test;
-#[path = "maven_golden_test.rs"]
+#[path = "maven/golden_test.rs"]
 mod maven_golden_test;
 #[path = "meson_golden_test.rs"]
 mod meson_golden_test;
