+++ src/parsers/maven/manifest_test.rs
@@ -1,8 +1,8 @@
 // SPDX-FileCopyrightText: Provenant contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use super::PackageParser;
-use super::maven::*;
+use super::super::PackageParser;
+use super::{MavenParser, manifest::*};
 use crate::models::DatasourceId;
 use crate::models::PackageType;
 use std::fs;
